# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: cmp是一个客户关系管理平台
  version: 1.0.0
  title: 客户关系管理平台
basePath: /cam
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /cus/customertrajectory/list:
    get:
      tags:
        - 客户轨迹管理
      summary: 客户轨迹列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: customerId
          description: 客户id
          in: query
          type: string
        - name: trajectoryName
          description: 轨迹名称
          in: query
          type: string
        - name: trajectoryType
          description: 轨迹类型
          in: query
          type: string
        - name: createCode
          description: 创建人代码
          in: query
          type: string
        - name: createName
          description: 创建人
          in: query
          type: string
        - name: startCreateTime
          description: 开始创建时间
          in: query
          type: string
          format: date
        - name: endCreateTime
          description: 结束创建时间
          in: query
          type: string
          format: date
        - name: startCreateTime
          description: 开始创建时间
          in: query
          type: string
          format: date
        - name: endCreateTime
          description: 结束创建时间
          in: query
          type: string
          format: date
      responses:
        '200':
          description: 返回客户轨迹列表
          schema:
            $ref: '#/definitions/CustomerTrajectoryEntityList'
  /cus/customertrajectory/info/{id}:
    get:
      tags:
        - 客户轨迹管理
      summary: 获取客户轨迹信息
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回客户轨迹信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/CustomerTrajectoryEntity'
  /cus/customertrajectory/save:
    post:
      tags:
        - 客户轨迹管理
      summary: 添加客户轨迹
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户轨迹对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CustomerTrajectoryEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/customertrajectory/update:
    post:
      tags:
        - 客户轨迹管理
      summary: 修改客户轨迹
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户轨迹对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CustomerTrajectoryEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/customertrajectory/delete:
    post:
      tags:
        - 客户轨迹管理
      summary: 删除客户轨迹
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户轨迹ID列表
          in: body
          type: array
          items:
            type: string
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /cus/customerbase/list:
    get:
      tags:
        - 客户信息管理
      summary: 客户信息列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: surname
          description: 姓
          in: query
          type: string
        - name: customerName
          description: 姓名
          in: query
          type: string
        - name: enName
          description: 英文名
          in: query
          type: string
        - name: beforeName
          description: 曾用名
          in: query
          type: string
        - name: nickname
          description: 昵称
          in: query
          type: string
        - name: idNumber
          description: 身份证号码
          in: query
          type: string
        - name: driverNumber
          description: 驾照号码
          in: query
          type: string
        - name: nationality
          description: 国籍
          in: query
          type: string
        - name: nation
          description: 民族
          in: query
          type: string
        - name: nativePlace
          description: 籍贯
          in: query
          type: string
        - name: sex
          description: 性别
          in: query
          type: string
        - name: age
          description: 年龄
          in: query
          type: integer
        - name: birthday
          description: 生日
          in: query
          type: string
        - name: height
          description: 身高
          in: query
          type: integer
        - name: weight
          description: 体重
          in: query
          type: integer
        - name: bloodType
          description: 血型
          in: query
          type: string
        - name: politicalAppearance
          description: 政治面貌
          in: query
          type: string
        - name: birth
          description: 生育
          in: query
          type: string
        - name: workingYears
          description: 工龄
          in: query
          type: integer
        - name: maritalStatus
          description: 婚姻状况
          in: query
          type: string
        - name: customerEducation
          description: 客户学历
          in: query
          type: string
        - name: headUrl
          description: 头像地址
          in: query
          type: string
        - name: company
          description: 公司
          in: query
          type: string
        - name: department
          description: 部门
          in: query
          type: string
        - name: duty
          description: 职务
          in: query
          type: string
        - name: belongCode
          description: 所属坐席
          in: query
          type: string
        - name: belongInstitutions
          description: 坐席所属机构
          in: query
          type: string
        - name: customerNotes
          description: 客户备注
          in: query
          type: string
        - name: isEffective
          description: 是否有效 0-否 1-是
          in: query
          type: string
        - name: createCode
          description: 创建人代码
          in: query
          type: string
        - name: createName
          description: 创建人
          in: query
          type: string
        - name: startCreateTime
          description: 开始创建时间
          in: query
          type: string
          format: date
        - name: endCreateTime
          description: 结束创建时间
          in: query
          type: string
          format: date
        - name: updateCode
          description: 修改人代码
          in: query
          type: string
        - name: updateName
          description: 修改人
          in: query
          type: string
        - name: startUpdateTime
          description: 开始修改时间
          in: query
          type: string
          format: date
        - name: endUpdateTime
          description: 结束修改时间
          in: query
          type: string
          format: date
        - name: isDelete
          description: 是否删除 0：否 1：是
          in: query
          type: string
      responses:
        '200':
          description: 返回客户信息列表
          schema:
            $ref: '#/definitions/CustomerBaseEntityList'
  /cus/customerbase/info/{id}:
    get:
      tags:
        - 客户信息管理
      summary: 获取客户信息信息
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回客户信息信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/CustomerBaseEntity'
  /cus/customerbase/save:
    post:
      tags:
        - 客户信息管理
      summary: 添加客户信息
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户信息对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CustomerBaseEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/customerbase/update:
    post:
      tags:
        - 客户信息管理
      summary: 修改客户信息
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户信息对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CustomerBaseEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/customerbase/delete:
    post:
      tags:
        - 客户信息管理
      summary: 删除客户信息
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户信息ID列表
          in: body
          type: array
          items:
            type: string
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  CustomerTrajectoryEntity:
    type: object
    properties:
      id:
        description: ID
        type: string
      customerId:
        description: 客户id
        type: string
      trajectoryName:
        description: 轨迹名称
        type: string
      trajectoryContent:
        description: 轨迹内容
        type: string
      trajectoryType:
        description: 轨迹类型
        type: string
      createCode:
        description: 创建人代码
        type: string
      createName:
        description: 创建人
        type: string
      createTime:
        description: 创建时间
        type: string
      updateCode:
        description: 修改人代码
        type: string
      updateName:
        description: 修改人
        type: string
      updateTime:
        description: 修改时间
        type: string
      customerTrajectorySonList:
        description: 客户轨迹子表列表
        type: array
        items:
          $ref: '#/definitions/CustomerTrajectorySonEntity'
  CustomerTrajectoryEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/CustomerTrajectoryEntity'
  CustomerTrajectorySonEntity:
    type: object
    properties:
      id:
        description: ID
        type: string
      trajectoryId:
        description: 轨迹id
        type: string
      businessId:
        description: 业务id
        type: string
      businessName:
        description: 业务名称
        type: string
      businessUrl:
        description: 业务url
        type: string
      createCode:
        description: 创建人代码
        type: string
      createName:
        description: 创建人
        type: string
      createTime:
        description: 创建时间
        type: string
      updateCode:
        description: 修改人代码
        type: string
      updateName:
        description: 修改人
        type: string
      updateTime:
        description: 修改时间
        type: string
...
  CustomerBaseEntity:
    type: object
    properties:
      id:
        description: ID
        type: string
      surname:
        description: 姓
        type: string
      customerName:
        description: 姓名
        type: string
      enName:
        description: 英文名
        type: string
      beforeName:
        description: 曾用名
        type: string
      nickname:
        description: 昵称
        type: string
      idNumber:
        description: 身份证号码
        type: string
      driverNumber:
        description: 驾照号码
        type: string
      nationality:
        description: 国籍
        type: string
      nation:
        description: 民族
        type: string
      nativePlace:
        description: 籍贯
        type: string
      sex:
        description: 性别
        type: string
      age:
        description: 年龄
        type: integer
        format: int32
      birthday:
        description: 生日
        type: string
      height:
        description: 身高
        type: integer
        format: int32
      weight:
        description: 体重
        type: integer
        format: int32
      bloodType:
        description: 血型
        type: string
      politicalAppearance:
        description: 政治面貌
        type: string
      birth:
        description: 生育
        type: string
      workingYears:
        description: 工龄
        type: integer
        format: int32
      maritalStatus:
        description: 婚姻状况
        type: string
      customerEducation:
        description: 客户学历
        type: string
      headUrl:
        description: 头像地址
        type: string
      company:
        description: 公司
        type: string
      department:
        description: 部门
        type: string
      duty:
        description: 职务
        type: string
      belongCode:
        description: 所属坐席
        type: string
      belongInstitutions:
        description: 坐席所属机构
        type: string
      customerNotes:
        description: 客户备注
        type: string
      isEffective:
        description: 是否有效 0-否 1-是
        type: string
      createCode:
        description: 创建人代码
        type: string
      createName:
        description: 创建人
        type: string
      createTime:
        description: 创建时间
        type: string
      updateCode:
        description: 修改人代码
        type: string
      updateName:
        description: 修改人
        type: string
      updateTime:
        description: 修改时间
        type: string
      isDelete:
        description: 是否删除 0：否 1：是
        type: string
  CustomerBaseEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/CustomerBaseEntity'
...